refactor: split ThorChainApi DTOs and helpers out of god file (#4244)
Move all 19 response DTOs to data/api/models/thorchain/, the cosmos tx
envelope to data/api/models/cosmos/, and the affiliate query-param logic
to a new ThorChainAffiliateHelper. ThorChainApi is now strictly an
HTTP-call surface that returns response models.

Also drops buildAffiliateParams from the interface (no external callers)
and makes the two denom-metadata helpers private on the impl.
