+++ data/src/main/kotlin/com/vultisig/wallet/data/api/models/thorchain/BlockNumber.kt
@@ -1,0 +1,5 @@
+package com.vultisig.wallet.data.api.models.thorchain
+
+import kotlinx.serialization.Serializable
+
+@Serializable data class BlockNumber(val thorchain: Long)
