+++ data/src/main/kotlin/com/vultisig/wallet/data/api/models/thorchain/ChurnEntry.kt
@@ -1,0 +1,10 @@
+package com.vultisig.wallet.data.api.models.thorchain
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ChurnEntry(
+    @SerialName("date") val date: String,
+    @SerialName("height") val height: String,
+)
