+++ data/src/main/kotlin/com/vultisig/wallet/data/api/models/thorchain/MidgardHealth.kt
@@ -1,0 +1,12 @@
+package com.vultisig.wallet.data.api.models.thorchain
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class MidgardHealth(val lastThorNode: HeightInfo) {
+    @Serializable
+    data class HeightInfo(
+        val height: Long,
+        val timestamp: Long, // seconds since epoch
+    )
+}
