+++ data/src/main/kotlin/com/vultisig/wallet/data/api/models/thorchain/TcyDistribution.kt
@@ -1,0 +1,6 @@
+package com.vultisig.wallet.data.api.models.thorchain
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class TcyDistribution(val block: String, val amount: String, val timestamp: String? = null)
