+++ data/src/main/kotlin/com/vultisig/wallet/data/api/models/thorchain/TcyModuleBalanceResponse.kt
@@ -1,0 +1,8 @@
+package com.vultisig.wallet.data.api.models.thorchain
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class TcyModuleBalanceResponse(val coins: List<ModuleCoin> = emptyList()) {
+    @Serializable data class ModuleCoin(val denom: String, val amount: String)
+}
