+++ data/src/main/kotlin/com/vultisig/wallet/data/api/models/thorchain/TcyStakeResponse.kt
@@ -1,0 +1,5 @@
+package com.vultisig.wallet.data.api.models.thorchain
+
+import kotlinx.serialization.Serializable
+
+@Serializable data class TcyStakeResponse(val amount: String? = null, val address: String = "")
