+++ data/src/main/kotlin/com/vultisig/wallet/data/api/models/thorchain/TcyStakersResponse.kt
@@ -1,0 +1,9 @@
+package com.vultisig.wallet.data.api.models.thorchain
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class TcyStakersResponse(@SerialName("tcy_stakers") val tcyStakers: List<TcyStaker>) {
+    @Serializable data class TcyStaker(val address: String, val amount: String)
+}
