+++ data/src/main/kotlin/com/vultisig/wallet/data/api/models/thorchain/ThorChainTransactionJson.kt
@@ -1,0 +1,11 @@
+package com.vultisig.wallet.data.api.models.thorchain
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ThorChainTransactionJson(
+    @SerialName("code") val code: Int?,
+    @SerialName("codespace") val codeSpace: String?,
+    @SerialName("rawLog") val rawLog: String,
+)
