+++ data/src/main/kotlin/com/vultisig/wallet/data/chains/helpers/ThorChainAffiliateHelper.kt
@@ -1,0 +1,46 @@
+package com.vultisig.wallet.data.chains.helpers
+
+object ThorChainAffiliateHelper {
+
+    fun buildAffiliateParams(referralCode: String, discountBps: Int): Map<String, String> {
+        val affiliateParams = mutableMapOf<String, String>()
+
+        // For ultimate clean referral, and return 0 bps affiliate
+        if (discountBps >= 50) {
+            affiliateParams["affiliate"] = THORChainSwaps.AFFILIATE_FEE_ADDRESS
+            affiliateParams["affiliate_bps"] = "0"
+
+            return affiliateParams
+        }
+
+        if (referralCode.isNotEmpty()) {
+            val affiliateFeeRateBp =
+                calculateBpsAfterDiscount(
+                    baseBps = THORChainSwaps.REFERRED_AFFILIATE_FEE_RATE_BP,
+                    discountBps = discountBps,
+                )
+
+            // Build nested affiliate with new thorchain structure
+            val affiliates = "$referralCode/${THORChainSwaps.AFFILIATE_FEE_ADDRESS}"
+            val affiliateBps = "${THORChainSwaps.REFERRED_USER_FEE_RATE_BP}/$affiliateFeeRateBp"
+
+            affiliateParams["affiliate"] = affiliates
+            affiliateParams["affiliate_bps"] = affiliateBps
+        } else {
+            val affiliateFeeRateBp =
+                calculateBpsAfterDiscount(
+                    baseBps = THORChainSwaps.AFFILIATE_FEE_RATE_BP,
+                    discountBps = discountBps,
+                )
+
+            affiliateParams["affiliate"] = THORChainSwaps.AFFILIATE_FEE_ADDRESS
+            affiliateParams["affiliate_bps"] = affiliateFeeRateBp.toString()
+        }
+
+        return affiliateParams
+    }
+
+    private fun calculateBpsAfterDiscount(baseBps: Int, discountBps: Int): Int {
+        return maxOf(0, baseBps - discountBps)
+    }
+}
